UInt - Size Warnings
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/branches/Release1.5@5828 6adb6e08-d915-0410-941f-83917bcadc18

+++ source/CHEMISTRY/AdvancedTheoreticalSpectrumGenerator.C
@@ -325,7 +325,7 @@ namespace OpenMS
             continue;
           }
           ion = &prefix;
-          ion_nr=i;
+          ion_nr=(UInt)i;
         }
         else if (residue == Residue::XIon || residue == Residue::YIon || residue == Residue::ZIon)
         {
@@ -334,7 +334,7 @@ namespace OpenMS
             continue;
           }
           ion = &suffix;
-          ion_nr=peptide.size()-i;
+          ion_nr=(UInt)(peptide.size()-i);
         }
         else
         {
